move loadSkillFiles to sandbox resource
